refactor: move createApi into it's own file
